refactor: 검색페이지 코드 분리
