+++ frontend/src/components/Search/NoRecentSearchResult.tsx
